Improve Cheevos & OSD, add path settings, Windows improvements
I added/prettified a bunch of OSD messages. Also prettified the code a little bit and removed some old debugging stuff that lingered. The Cheevos login now uses the proper functions as well, and token expiration is now handled. Also, the controller and such now initializes on Windows (previously that entire config area was locked behind whether it was running on Xbox). Additionally, due to issues with launching browsers on UWP for some reason, the buttons on the About page are now disabled on UWP.